Started to add oversampling receiver.

+++ tools/gr-marmote/lib/pn_spreader_f_impl.h
@@ -1,17 +1,17 @@
 /* -*- c++ -*- */
-/* 
+/*
  * Copyright 2013 <+YOU OR YOUR COMPANY+>.
- * 
+ *
  * This is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 3, or (at your option)
  * any later version.
- * 
+ *
  * This software is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this software; see the file COPYING.  If not, write to
  * the Free Software Foundation, Inc., 51 Franklin Street,
@@ -37,7 +37,7 @@ namespace gr {
       unsigned int d_spread_factor;
       unsigned int d_preamble_len;
 
-      float d_chip_buf[MAX_CHIP_BUF_LEN];
+      uint8_t d_chip_buf[MAX_CHIP_BUF_LEN];
       int d_chip_offset;
       int d_chip_len;
